Avoid g++ warnings due to non-excplicit type casts
git-svn-id: https://deskvox.svn.sourceforge.net/svnroot/deskvox/trunk@370 04231f92-3938-0410-9931-e931ac552b4f

+++ virvo/src/vvoffscreenbuffer.cpp
@@ -148,8 +148,8 @@ BufferPrecision vvOffscreenBuffer::getPrecision() const
 
 void vvOffscreenBuffer::doScale()
 {
-  _bufferWidth = (float)_viewportWidth * _scale;
-  _bufferHeight = (float)_viewportHeight * _scale;
+  _bufferWidth = (int)((float)_viewportWidth * _scale);
+  _bufferHeight = (int)((float)_viewportHeight * _scale);
 }
 
 void vvOffscreenBuffer::update()
